00:47 20/09/2022

+++ modules/handle/delete-address.php
@@ -1,12 +1,15 @@
 <?php
     include_once './modules/handle/connect-database.php';
-    $id=$_GET['id'];
-    if($connect){
-        $sql='delete from orderaddress where id='.$id;
-        if(mysqli_query($connect,$sql)){
-            echo '<a href="dia-chi-nhan-hang"></a>';
-        }
+    if(isset($_GET['id'])){
+        if($connect){
+            $sql='delete from orderaddress where id='.$id;
+            if(mysqli_query($connect,$sql)){
+                echo '<a href="dia-chi-nhan-hang"></a>';
+            }else echo '<a href="dia-chi-nhan-hang"></a>';
 
+        }
+    }else{
+        header("location:trang-chu");
     }
 ?>
